+++ src/main/java/eu/openanalytics/phaedra/phaedra2resultdataservice/model/ResultData.java
@@ -17,7 +17,7 @@ public class ResultData {
 
     Long featureId;
 
-    double[] values;
+    float[] values;
 
     StatusCode statusCode;
